로그인한 사용자에 대한 게시글 작성/수정/삭제 구현

+++ src/main/java/com/sparta/blog/service/BlogService.java
@@ -2,6 +2,7 @@ package com.sparta.blog.service;
 
 import com.sparta.blog.dto.BlogRequestDto;
 import com.sparta.blog.dto.BlogResponseDto;
+import com.sparta.blog.entity.User;
 import com.sparta.blog.repository.BlogRepository;
 import com.sparta.blog.entity.Blog;
 import org.springframework.stereotype.Component;
@@ -21,32 +22,42 @@ public class BlogService {
         return blogRepository.findAllByOrderByModifiedAtDesc().stream().map(BlogResponseDto::new).toList();
     }
 
-    public BlogResponseDto createBlog(BlogRequestDto requestDto) {
-        Blog blog = new Blog(requestDto);
+    public BlogResponseDto createBlog(BlogRequestDto requestDto, User user) {
+        Blog blog = new Blog(requestDto, user);
         Blog saveBlog = blogRepository.save(blog);
         BlogResponseDto blogResponseDto = new BlogResponseDto(blog);
         return blogResponseDto;
     }
 
     @Transactional
-    public Long updateBlog(Long id, BlogRequestDto requestDto) {
+    public BlogResponseDto updateBlog(Long id, BlogRequestDto requestDto, User user) {
         // 해당 게시글이 DB에 존재하는지 확인
         Blog blog = findBlog(id);
+        Blog updatedBlog;
         if (blog != null) {
-            blog.update(requestDto);
-
-            return id;
+            if(user.getUsername().equals(blog.getUsername())) {
+                blog.update(requestDto, user);
+                updatedBlog = blog;
+            }
+            else{
+                throw new IllegalArgumentException("로그인한 작성자의 게시물이 아닙니다.");
+            }
+            return new BlogResponseDto(updatedBlog);
         } else {
-            throw new IllegalArgumentException("선택한 게시글은 존재하지 않습니다.");
+            throw new IllegalArgumentException("선택한 게시글은 존재하지 않습니다. 수정할 수 없습니다.");
         }
     }
 
-    public Long deleteBlog(Long id) {
+    public Long deleteBlog(Long id, User user) {
         // 해당 게시글이 DB에 존재하는지 확인
         Blog blog =findBlog(id);
         if (blog != null) {
-            // memo 삭제
-            blogRepository.delete(blog);
+            if(user.getUsername().equals(blog.getUsername())) {
+                blogRepository.delete(blog);
+            }else{
+                throw new IllegalArgumentException("로그인한 작성자의 게시물이 아닙니다. 삭제할 수 없습니다.");
+            }
+
             return id;
         } else {
             throw new IllegalArgumentException("선택한 게시글은 존재하지 않습니다.");
@@ -66,7 +77,7 @@ public class BlogService {
 
     private Blog findBlog(Long id){
         return blogRepository.findById(id).orElseThrow(()->
-                new IllegalArgumentException("선택한 메모는 존재하지 않습니다.")
+                new IllegalArgumentException("선택한 게시글은 존재하지 않습니다.")
         );
     }
 }
